django/requirements.txt
# Core dependencies
langchain==0.1.16
langchain-community>=0.0.20
langchain-openai>=0.1.3
openapi-python-client>=0.15.0
requests>=2.31.0
atlassian-python-api==3.41.16
psycopg==3.2.3
# Django REST framework for API views
djangorestframework>=3.15.1
# For type checking and linting (optional, but recommended)
types-requests